Move version in package name
